move filewalker
